+++ websauna/compat/typing.py
@@ -1,6 +1,3 @@
 import sys
 
-try:
-    from typing import  *  # pragma: no cover
-except ImportError:  # pragma: no cover
-    from backports.typing import *  # pragma: no cover
+from typing import  *  # pragma: no cover